+++ ComponentImpl/src/main/java/com/xiaojinzi/component/impl/RouterFragment.java
@@ -30,7 +30,6 @@ import java.util.Set;
 public final class RouterFragment extends Fragment {
 
     public RouterFragment() {
-        System.out.println("123123");
     }
 
     @NonNull